Add files via upload
I added the attenuation of the audio signal with distance (assumed point source so -6 dBFS with doubling of distance). Modified the recordings (normalization), assumed 0 dBFS at 1 m distance for speech signal and -2 dBFS for the noise signal (it would be too loud otherwise for a realistic analysis.) 

Calculation of attenuation is added to the compute_delay_and_attenuation function (formerly compute_delay).